extract ControlValueAccessor
