+++ src/frontend/components/common/helper/InputValueControlAccessor.ts
@@ -1,0 +1,50 @@
+import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
+import {forwardRef} from '@angular/core';
+
+const noop = () => {
+};
+
+export function ProvideValueAccessor(component: any) {
+    return {
+        provide: NG_VALUE_ACCESSOR,
+        useExisting: forwardRef(() => component),
+        multi: true
+    };
+}
+
+export class InputValueControlAccessor implements ControlValueAccessor {
+
+    private innerValue: any = '';
+
+    private onTouchedCallback: () => void = noop;
+    private onChangeCallback: (_: any) => void = noop;
+
+    get value(): any {
+        return this.innerValue;
+    };
+
+    set value(v: any) {
+        if (v !== this.innerValue) {
+            this.innerValue = v;
+            this.onChangeCallback(v);
+        }
+    }
+
+    onBlur() {
+        this.onTouchedCallback();
+    }
+
+    writeValue(value: any) {
+        if (value !== this.innerValue) {
+            this.innerValue = value;
+        }
+    }
+
+    registerOnChange(fn: any) {
+        this.onChangeCallback = fn;
+    }
+
+    registerOnTouched(fn: any) {
+        this.onTouchedCallback = fn;
+    }
+}
