+++ src/frontend/components/elements/checkbox/checkbox.component.ts
@@ -1,15 +1,5 @@
-import { Component, Input, forwardRef } from '@angular/core';
-import { NG_VALUE_ACCESSOR, ControlValueAccessor } from '@angular/forms';
-
-
-const noop = () => {
-};
-
-export const CUSTOM_CHECKBOX_CONTROL_VALUE_ACCESSOR: any = {
-    provide: NG_VALUE_ACCESSOR,
-    useExisting: forwardRef(() => CheckboxComponent),
-    multi: true
-};
+import {Component, Input} from '@angular/core';
+import {InputValueControlAccessor, ProvideValueAccessor} from '../../common/helper/InputValueControlAccessor';
 
 
 @Component({
@@ -17,9 +7,9 @@ export const CUSTOM_CHECKBOX_CONTROL_VALUE_ACCESSOR: any = {
     selector: 'checkbox',
     templateUrl: 'checkbox.component.html',
     styleUrls: ['checkbox.component.css'],
-    providers: [CUSTOM_CHECKBOX_CONTROL_VALUE_ACCESSOR]
+    providers: [ProvideValueAccessor(CheckboxComponent)]
 })
-export class CheckboxComponent implements ControlValueAccessor {
+export class CheckboxComponent extends InputValueControlAccessor {
 
     @Input()
     id: string;
@@ -28,51 +18,7 @@ export class CheckboxComponent implements ControlValueAccessor {
     @Input()
     label: string;
 
-
     constructor() {
-    }
-
-
-    //The internal data model
-    private innerValue: any = '';
-
-    //Placeholders for the callbacks which are later providesd
-    //by the Control Value Accessor
-    private onTouchedCallback: () => void = noop;
-    private onChangeCallback: (_: any) => void = noop;
-
-    //get accessor
-    get value(): any {
-        return this.innerValue;
-    };
-
-    //set accessor including call the onchange callback
-    set value(v: any) {
-        if (v !== this.innerValue) {
-            this.innerValue = v;
-            this.onChangeCallback(v);
-        }
-    }
-
-    //Set touched on blur
-    onBlur() {
-        this.onTouchedCallback();
-    }
-
-    //From ControlValueAccessor interface
-    writeValue(value: any) {
-        if (value !== this.innerValue) {
-            this.innerValue = value;
-        }
-    }
-
-    //From ControlValueAccessor interface
-    registerOnChange(fn: any) {
-        this.onChangeCallback = fn;
-    }
-
-    //From ControlValueAccessor interface
-    registerOnTouched(fn: any) {
-        this.onTouchedCallback = fn;
+        super();
     }
 }
